Fix last commit

+++ Editor/Core/Ink Settings/InkSettings.cs
@@ -109,10 +109,13 @@ namespace Ink.UnityIntegration {
 		// Deletes the persistent version of this asset that we used to use prior to 0.9.71
 		void OnEnable () {
 			if(!Application.isPlaying && EditorUtility.IsPersistent(this)) {
-				if(_Instance == this) _Instance = null;
-				AssetDatabase.DeleteAsset(AssetDatabase.GetAssetPath(this));
-				AssetDatabase.Refresh();
-				return;
+				var path = AssetDatabase.GetAssetPath(this);
+				if(path != null) {
+					if(_Instance == this) _Instance = null;
+					AssetDatabase.DeleteAsset(AssetDatabase.GetAssetPath(this));
+					AssetDatabase.Refresh();
+					return;
+				}
 			}
 		}
